requirements.txt
python-dotenv
boto3
pg8000